refactor: create a dir(models) and separate each model into one file
